Tests - Split tests up before plan change
